my_project_new_features_branch_s9 fix comment

+++ accounts-service/src/main/java/ru/yandex/practicum/accounts/dto/AccountResponseDto.java
@@ -1,6 +1,7 @@
 package ru.yandex.practicum.accounts.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record AccountResponseDto(
@@ -8,6 +9,6 @@ public record AccountResponseDto(
         String name,
         @JsonFormat(pattern = "yyyy-MM-dd")
         LocalDate birthdate,
-        Integer sum
+        BigDecimal sum
 ) {
 }
