+++ cash-service/src/main/java/ru/yandex/practicum/cash/dto/AccountResponseDto.java
@@ -1,9 +1,11 @@
 package ru.yandex.practicum.cash.dto;
 
+import java.math.BigDecimal;
+
 public record AccountResponseDto(
         String login,
         String name,
         String birthdate,
-        Integer sum
+        BigDecimal sum
 ) {
 }
