+++ cash-service/src/main/java/ru/yandex/practicum/cash/dto/CashRequestDto.java
@@ -3,6 +3,7 @@ package ru.yandex.practicum.cash.dto;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 public record CashRequestDto(
         @NotBlank(message = "Login is required")
@@ -10,7 +11,7 @@ public record CashRequestDto(
 
         @NotNull(message = "Value is required")
         @Min(value = 1, message = "Amount must be positive")
-        Integer value,
+        BigDecimal value,
 
         @NotNull(message = "Action is required")
         CashAction action
