+++ transfer-service/src/main/java/ru/yandex/practicum/transfer/dto/AccountResponseDto.java
@@ -1,9 +1,11 @@
 package ru.yandex.practicum.transfer.dto;
 
+import java.math.BigDecimal;
+
 public record AccountResponseDto(
         String login,
         String name,
         String birthdate,
-        Integer sum
+        BigDecimal sum
 ) {
 }
